+++ public/src/components/channelManagement/bannerTests/bannerTestEditor.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {Region} from '../../../utils/models';
-import {ArticlesViewedSettings, UserCohort} from "../helpers/shared";
+import {articleCountTemplate, ArticlesViewedSettings, TestEditorState, UserCohort} from "../helpers/shared";
 import {
   createStyles,
   Switch,
@@ -13,14 +13,12 @@ import {onFieldValidationChange, isNumber} from '../helpers/validation';
 import ButtonWithConfirmationPopup from '../helpers/buttonWithConfirmationPopup';
 import DeleteSweepIcon from '@material-ui/icons/DeleteSweep';
 import ArchiveIcon from '@material-ui/icons/Archive';
-import {TestEditorState} from "../epicTests/epicTestEditor";
 import {BannerTest, BannerVariant} from "./bannerTestsForm";
-import TargetRegionsSelector from "../epicTests/targetRegionsSelector";
+import TargetRegionsSelector from "../targetRegionsSelector";
 import ArticlesViewedEditor, {defaultArticlesViewedSettings} from "../articlesViewedEditor";
-import {articleCountTemplate} from "../epicTests/epicTestVariantEditor";
-import NewNameCreator from "../epicTests/newNameCreator";
+import NewNameCreator from "../newNameCreator";
 import BannerTestVariantsList from "./bannerTestVariantsList";
-import UserCohortSelector from "../epicTests/userCohortSelector";
+import UserCohortSelector from "../userCohortSelector";
 import EditableTextField from "../helpers/editableTextField"
