+++ public/src/components/channelManagement/helpers/shared.tsx
@@ -1,3 +1,9 @@
+import {ValidationStatus} from "./validation";
+
+export const currencyTemplate = "%%CURRENCY_SYMBOL%%";
+export const countryNameTemplate = "%%COUNTRY_NAME%%";
+export const articleCountTemplate = "%%ARTICLE_COUNT%%";
+
 export interface LockStatus {
   locked: boolean,
   email?: string,
@@ -16,6 +22,12 @@ export interface Cta {
   baseUrl?: string
 }
 
+export const defaultCta = {
+  text: "Support The Guardian",
+  baseUrl: "https://support.theguardian.com/contribute"
+};
+
+
 export enum UserCohort {
   Everyone = 'Everyone',
   AllExistingSupporters = 'AllExistingSupporters',
@@ -32,4 +44,8 @@ export interface ArticlesViewedSettings {
   minViews: number | null,
   maxViews: number | null,
   periodInWeeks: number,
+}
+
+export interface TestEditorState {
+    validationStatus: ValidationStatus
 }
